Module google_webrisk1::api

Structs

GoogleCloudWebriskV1ComputeThreatListDiffResponse
There is no detailed description.

GoogleCloudWebriskV1ComputeThreatListDiffResponseChecksum
The expected state of a client's local database.

GoogleCloudWebriskV1RawHashes
The uncompressed threat entries in hash format. Hashes can be anywhere from 4 to 32 bytes in size. A large majority are 4 bytes, but some hashes are lengthened if they collide with the hash of a popular URI. Used for sending ThreatEntryAdditons to clients that do not support compression, or when sending non-4-byte hashes to clients that do support compression.

GoogleCloudWebriskV1RawIndices
A set of raw indices to remove from a local list.

GoogleCloudWebriskV1RiceDeltaEncoding
The Rice-Golomb encoded data. Used for sending compressed 4-byte hashes or compressed removal indices.

GoogleCloudWebriskV1SearchHashesResponse
There is no detailed description.

GoogleCloudWebriskV1SearchHashesResponseThreatHash
Contains threat information on a matching hash.

GoogleCloudWebriskV1SearchUrisResponse
There is no detailed description.

GoogleCloudWebriskV1SearchUrisResponseThreatUri
Contains threat information on a matching uri.

GoogleCloudWebriskV1Submission
Wraps a URI that might be displaying malicious content.

GoogleCloudWebriskV1SubmitUriRequest
Request to send a potentially malicious URI to WebRisk.

GoogleCloudWebriskV1ThreatEntryAdditions
Contains the set of entries to add to a local database. May contain a combination of compressed and raw data in a single response.

GoogleCloudWebriskV1ThreatEntryRemovals
Contains the set of entries to remove from a local database.

GoogleLongrunningCancelOperationRequest
The request message for Operations.CancelOperation.

GoogleLongrunningListOperationsResponse
The response message for Operations.ListOperations.

GoogleLongrunningOperation
This resource represents a long-running operation that is the result of a network API call.

GoogleProtobufEmpty
A generic empty message that you can re-use to avoid defining duplicated empty messages in your APIs. A typical example is to use it as the request or the response type of an API method. For instance: service Foo { rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty); } The JSON representation for Empty is empty JSON object {}.

GoogleRpcStatus
The Status type defines a logical error model that is suitable for different programming environments, including REST APIs and RPC APIs. It is used by gRPC. Each Status message contains three pieces of data: error code, error message, and error details. You can find out more about this error model and how to work with it in the API Design Guide.

HasheMethods
A builder providing access to all methods supported on hashe resources. It is not used directly, but through the WebRisk hub.

HasheSearchCall
Gets the full hashes that match the requested hash prefix. This is used after a hash prefix is looked up in a threatList and there is a match. The client side threatList only holds partial hashes so the client must query this method to determine if there is a full hash match of a threat.

ProjectMethods
A builder providing access to all methods supported on project resources. It is not used directly, but through the WebRisk hub.

ProjectOperationCancelCall
Starts asynchronous cancellation on a long-running operation. The server makes a best effort to cancel the operation, but success is not guaranteed. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED. Clients can use Operations.GetOperation or other methods to check whether the cancellation succeeded or whether the operation completed despite cancellation. On successful cancellation, the operation is not deleted; instead, it becomes an operation with an Operation.error value with a google.rpc.Status.code of 1, corresponding to Code.CANCELLED.

ProjectOperationDeleteCall
Deletes a long-running operation. This method indicates that the client is no longer interested in the operation result. It does not cancel the operation. If the server doesn't support this method, it returns google.rpc.Code.UNIMPLEMENTED.

ProjectOperationGetCall
Gets the latest state of a long-running operation. Clients can use this method to poll the operation result at intervals as recommended by the API service.

ProjectOperationListCall
Lists operations that match the specified filter in the request. If the server doesn't support this method, it returns UNIMPLEMENTED. NOTE: the name binding allows API services to override the binding to use different resource name schemes, such as users/*/operations. To override the binding, API services can add a binding such as "/v1/{name=users/*}/operations" to their service configuration. For backwards compatibility, the default name includes the operations collection id, however overriding users must ensure the name binding is the parent resource, without the operations collection id.

ProjectSubmissionCreateCall
Creates a Submission of a URI suspected of containing phishing content to be reviewed. If the result verifies the existence of malicious phishing content, the site will be added to the Google's Social Engineering lists in order to protect users that could get exposed to this threat in the future. Only allowlisted projects can use this method during Early Access. Please reach out to Sales or your customer engineer to obtain access.

ProjectUriSubmitCall
Submits a URI suspected of containing malicious content to be reviewed. Returns a google.longrunning.Operation which, once the review is complete, is updated with its result. You can use the [Pub/Sub API] (https://cloud.google.com/pubsub) to receive notifications for the returned Operation. If the result verifies the existence of malicious content, the site will be added to the [Google's Social Engineering lists] (https://support.google.com/webmasters/answer/6350487/) in order to protect users that could get exposed to this threat in the future. Only allowlisted projects can use this method during Early Access. Please reach out to Sales or your customer engineer to obtain access.

ThreatListComputeDiffCall
Gets the most recent threat list diffs. These diffs should be applied to a local database of hashes to keep it up-to-date. If the local database is empty or excessively out-of-date, a complete snapshot of the database will be returned. This Method only updates a single ThreatList at a time. To update multiple ThreatList databases, this method needs to be called once for each list.

ThreatListMethods
A builder providing access to all methods supported on threatList resources. It is not used directly, but through the WebRisk hub.

UriMethods
A builder providing access to all methods supported on uri resources. It is not used directly, but through the WebRisk hub.

UriSearchCall
This method is used to check whether a URI is on a given threatList. Multiple threatLists may be searched in a single query. The response will list all requested threatLists the URI was found to match. If the URI is not found on any of the requested ThreatList an empty response will be returned.

WebRisk
Central instance to access all WebRisk related resource activities

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.